pyproject.toml
[project]
name = "cv25"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "daphne>=4.1.2",
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "jupyter>=1.1.1",
    "numpy>=2.3.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pdf2image",
    "docx>=0.2.4",
    "python-docx>=1.2.0",
    "pillow>=11.3.0",
    "playwright>=1.57.0",
    "easyocr>=1.7.2",
    "opencv-python>=4.13.0.90",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-playwright>=0.4.0",
    "pylint>=3.0.0",
    "mypy>=1.8.0",
    "types-requests",
    "black>=26.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning:PyPDF2.*:",
]
markers = [
    "e2e: End-to-end tests using Playwright (deselect with '-m \"not e2e\"')",
    "slow: Tests that take a long time to run",
]
